+++ lib/qc/modules/qc_tester/lob_devices/screens/product_list_screen.g.dart
@@ -1,0 +1,25 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'product_list_screen.dart';
+
+// **************************************************************************
+// Generator: PageSchemaGen
+// **************************************************************************
+
+dynamic schema() => {
+      //#adminscreen
+      "type": "@@page",
+      "key": "QC_product_list_screen",
+      "pageGroup": "QC Product List Page",
+      "title": "Product List Screen",
+      "isActive": true,
+      "params": [
+        {"key": "dbr", "value": null},
+        {"key": "bid", "value": null},
+        {"key": "cid", "value": null},
+        {"key": "cat", "value": null},
+        {"key": "ops", "value": null}
+      ]
+
+      //#adminscreen
+    };