+++ lib/qc/modules/qc_tester/lob_devices/widgets/lob_device_scanner_widget.dart
@@ -1,32 +1,52 @@
+import 'package:core/core.dart';
 import 'package:core_widgets/core_widgets.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_trc/qc/modules/qc_tester/calculator/models/calculator_data_holder_model.dart';
+import 'package:flutter_trc/qc/modules/qc_tester/calculator/screens/calculation_screen.dart';
 import 'package:flutter_trc/qc/modules/qc_tester/lob_devices/providers/lob_device_scanner_provider.dart';
 import 'package:flutter_trc/qc/modules/qc_tester/lob_devices/resources/lob_product_list_response.dart';
+import 'package:flutter_trc/qc/modules/qc_tester/lob_devices/resources/variant_list_response.dart';
+import 'package:flutter_trc/qc/modules/qc_tester/lob_devices/screens/product_list_screen.dart';
 import 'package:flutter_trc/qc/modules/qc_tester/lob_devices/widgets/lob_device_detail_widget.dart';
-import 'package:flutter_trc/qc/modules/qc_tester/lob_devices/widgets/product_list_widget.dart';
 import 'package:flutter_trc/src/common/utils/csh_ml_scanner_util.dart';
-import 'package:provider/provider.dart';
 
 class LobDeviceScannerWidget extends StatelessWidget {
   const LobDeviceScannerWidget({super.key});
 
-  _oSearched(BuildContext context, bool isManual, int selectedCategoryId) {
+  _oSearched(BuildContext context, int selectedCategoryId, int selectedBrandId) {
+    var provider = LobDeviceScannerProvider.of(context, listen: false);
+    ProductListScreen.navigateTo(
+      context,
+      provider.deviceBarcode!,
+      selectedCategoryId,
+      selectedBrandId,
+      provider.getCategoryList() ?? [],
+      (productItem, variantItem) {
+        _getCalculator(context, productItem, selectedCategoryId, variantItem: variantItem);
+      },
+    );
+  }
+
+  _getCalculator(BuildContext context, LobProductListData item, int selectedCategoryId,
+      {VariantListData? variantItem}) {
     var provider = LobDeviceScannerProvider.of(context, listen: false);
     CshLoading().showLoading(context);
-    provider
-        .getProductsList(
-      provider.deviceBarcode!,
-      provider.deviceDetails?.imei1,
-      provider.deviceDetails?.serialNo,
-      isManual,
-      selectedCategoryId,
-    )
-        .then((value) {
+    provider.getLobCalculator(item.productMasterId, item.productId, selectedCategoryId, variantItem).then(
+        (calculatorResponse) {
       CshLoading().hideLoading(context);
-      _showProductListDialog(context, value!, provider.deviceBarcode!, provider, selectedCategoryId);
+      calculatorResponse?.brandId ??= item.brandId;
+      CalculatorDataHolderModel().startCalculatorJourney(
+        calculatorResponse,
+        provider.deviceBarcode,
+        deviceType: DeviceType.lob_device,
+        selectedCategoryId: selectedCategoryId,
+        variantData: variantItem,
+      );
+      Navigator.popUntil(context, (route) => route is PageRoute); // Dismiss all dialog
+      Navigator.pushReplacementNamed(context, CalculationScreen.route);
     }, onError: (error) {
       CshLoading().hideLoading(context);
-      CshSnackBar.error(context: context, message: error, snackBarPosition: SnackBarPosition.TOP);
+      CshSnackBar.error(context: context, message: ApiErrorHelper.getErrorMessage(error).toString());
     });
   }
 
@@ -65,24 +85,11 @@ class LobDeviceScannerWidget extends StatelessWidget {
       return LobDeviceDetailWidget(
         scannedData: provider.deviceBarcode!,
         deviceDetails: provider.deviceDetails,
-        onSearchClicked: (isManual, selectedCategoryId) {
-          _oSearched(context, isManual, selectedCategoryId);
+        onSearchClicked: (selectedBrandId, selectedCategoryId) {
+          _oSearched(context, selectedCategoryId, selectedBrandId);
         },
       );
     }
     return const SizedBox.shrink();
-  }
-
-  void _showProductListDialog(BuildContext context, List<LobProductListData> productList, String deviceBarcode,
-      LobDeviceScannerProvider provider, int selectedCategoryId) {
-    showCshBottomSheet(
-      context: context,
-      isDismissible: false,
-      isScrollControlled: true,
-      child: ChangeNotifierProvider.value(
-        value: provider,
-        child: ProductListWidget(productList, deviceBarcode, selectedCategoryId),
-      ),
-    );
   }
 }