+++ lib/src/app_builder/app_builder_groups/qc_groups.dart
@@ -34,7 +34,9 @@ enum QcComponentGroup with AbsComponentGroup {
   qcDeviceDetailsComponentKey("QC Device Details Component"),
   qcStorageDeviceListComponentKey("QC Storage Device List Component"),
   qcImeiValidatorComponentKey("QC Imei Validator Component"),
-  qcD2cVideoComponentKey("QC D2C Video Component");
+  qcD2cVideoComponentKey("QC D2C Video Component"),
+  qcProductListComponentKey("QC Product List Component"),
+  qcVariantListComponentKey("QC Variant List Component");
 
   @override
   final String value;
@@ -77,7 +79,9 @@ enum QcPageGroup with AbsPageGroup {
   qcDeviceDetailsPageKey("QC Device Details"),
   qcStorageDeviceListPageKey("QC Storage Device List"),
   qcImeiValidatorPageKey("QC Imei Validator"),
-  qcD2cVideoPageKey("QC D2C Video");
+  qcD2cVideoPageKey("QC D2C Video"),
+  qcProductListPageKey("QC Product List Page"),
+  qcVariantListPageKey("QC Variant List Page");
 
   @override
   final String value;